chore(deps): update rusqlite from 0.35 to 0.40.1
Rusqlite 0.40.1 removed FromSql/ToSql implementations for u64.
Updated FileMetadata.size from u64 to i64 to match the new API.
Updated IndexService to cast document.size to i64 when creating
FileMetadata records.

+++ crates/glintindex-core/src/index/service.rs
@@ -247,7 +247,7 @@ impl DocumentIndexer for IndexService {
         if let Some(ref repo) = self.metadata {
             let meta = FileMetadata {
                 path: document.path.to_string_lossy().to_string(),
-                size: document.size,
+                size: document.size as i64,
                 modified: document
                     .modified
                     .duration_since(std::time::UNIX_EPOCH)
